+++ tests/unit/containers/edit-profile/password-security-tab/ChangePasswordModal.spec.jsx
@@ -1,0 +1,233 @@
+import { vi } from 'vitest'
+import { screen, fireEvent, waitFor } from '@testing-library/react'
+import {
+  renderWithProviders,
+  mockAxiosClient,
+  TestSnackbar
+} from '~tests/test-utils'
+import PasswordSecurityTab from '~/containers/edit-profile/password-security-tab/PasswordSecurityTab'
+import { AuthService } from '~/services/auth-service'
+
+import { URLs } from '~/constants/request'
+
+const userDataMock = {
+  _id: 123456
+}
+
+vi.mock('~/services/auth-service', () => ({
+  AuthService: {
+    changePassword: vi.fn()
+  }
+}))
+
+const handleSubmit = vi.fn()
+
+const changeInputValue = (label, value) => {
+  fireEvent.change(label, { target: { value } })
+}
+
+const setupChangePasswordButton = () => {
+  const changePasswordButton = screen.getByText(
+    /editProfilePage.profile.passwordSecurityTab.changePassword/i
+  )
+  return {
+    changePasswordButton
+  }
+}
+const setupChangePasswordFields = () => {
+  const currentPasswordInput = screen.getByLabelText(
+    /editProfilePage.profile.passwordSecurityTab.currentPassword/i
+  )
+  const passwordInput = screen.getByLabelText(
+    /editProfilePage.profile.passwordSecurityTab.newPassword/i
+  )
+  const confirmPasswordInput = screen.getByLabelText(
+    /editProfilePage.profile.passwordSecurityTab.retypePassword/i
+  )
+  const saveButton = screen.getByText(
+    /editProfilePage.profile.passwordSecurityTab.savePassword/i
+  )
+
+  return {
+    currentPasswordInput,
+    passwordInput,
+    confirmPasswordInput,
+    saveButton
+  }
+}
+
+const renderWithMockData = () => {
+  renderWithProviders(
+    <TestSnackbar>
+      <PasswordSecurityTab />
+    </TestSnackbar>,
+    {
+      preloadedState: {
+        appMain: {
+          userId: userDataMock._id,
+          userStatus: 'active'
+        }
+      }
+    }
+  )
+}
+
+describe('PasswordSecurityTab', () => {
+  beforeEach(() => {
+    renderWithMockData()
+  })
+
+  it('should save data after positive response', async () => {
+    mockAxiosClient
+      .onPatch(`${URLs.auth.changePassword}/${userDataMock._id}`)
+      .reply(200)
+
+    const { changePasswordButton } = setupChangePasswordButton()
+
+    fireEvent.click(changePasswordButton)
+
+    const {
+      currentPasswordInput,
+      passwordInput,
+      confirmPasswordInput,
+      saveButton
+    } = setupChangePasswordFields()
+
+    fireEvent.change(currentPasswordInput, {
+      target: { value: '12345qwert' }
+    })
+    fireEvent.change(passwordInput, {
+      target: { value: '12345qwertY' }
+    })
+    fireEvent.change(confirmPasswordInput, {
+      target: { value: '12345qwertY' }
+    })
+
+    await waitFor(() => {
+      fireEvent.click(saveButton)
+    })
+
+    const confirmButton = screen.getByText('common.yes')
+    fireEvent.click(confirmButton)
+
+    await waitFor(() => {
+      expect(AuthService.changePassword).toHaveBeenCalledWith(
+        userDataMock._id,
+        {
+          password: '12345qwertY',
+          currentPassword: '12345qwert'
+        }
+      )
+    })
+  })
+
+  it('should not save data after negative response', async () => {
+    mockAxiosClient
+      .onPatch(`${URLs.auth.changePassword}/${userDataMock}`)
+      .reply(400, {
+        message: 'new password cannot be the same as the current one'
+      })
+    const { changePasswordButton } = setupChangePasswordButton()
+
+    fireEvent.click(changePasswordButton)
+
+    const {
+      currentPasswordInput,
+      passwordInput,
+      confirmPasswordInput,
+      saveButton
+    } = setupChangePasswordFields()
+
+    fireEvent.change(currentPasswordInput, {
+      target: { value: '12345qwertY' }
+    })
+    fireEvent.change(passwordInput, {
+      target: { value: '12345qwertY' }
+    })
+    fireEvent.change(confirmPasswordInput, {
+      target: { value: '12345qwertY' }
+    })
+
+    await waitFor(() => {
+      fireEvent.click(saveButton)
+    })
+    expect(handleSubmit).not.toHaveBeenCalled()
+  })
+
+  it('should do not save empty fields', async () => {
+    const { changePasswordButton } = setupChangePasswordButton()
+
+    fireEvent.click(changePasswordButton)
+    const {
+      currentPasswordInput,
+      passwordInput,
+      confirmPasswordInput,
+      saveButton
+    } = setupChangePasswordFields()
+
+    fireEvent.change(currentPasswordInput, {
+      target: { value: '' }
+    })
+    fireEvent.change(passwordInput, {
+      target: { value: '' }
+    })
+    fireEvent.change(confirmPasswordInput, {
+      target: { value: '' }
+    })
+
+    fireEvent.click(saveButton)
+
+    expect(handleSubmit).not.toHaveBeenCalled()
+
+    expect(currentPasswordInput).toHaveValue('')
+    expect(passwordInput).toHaveValue('')
+    expect(confirmPasswordInput).toHaveValue('')
+  })
+
+  it('should show visibility icon', async () => {
+    const { changePasswordButton } = setupChangePasswordButton()
+
+    fireEvent.click(changePasswordButton)
+    const visibilityOffIcons = screen.getAllByTestId('VisibilityOffIcon')
+    const visibilityOffIcon = visibilityOffIcons[0]
+    fireEvent.click(visibilityOffIcon)
+
+    const visibilityIcons = screen.getAllByTestId('VisibilityIcon')
+    const visibilityIcon = visibilityIcons[0]
+    await waitFor(() => {
+      expect(visibilityIcon).toBeInTheDocument()
+      expect(visibilityOffIcon).not.toBeInTheDocument()
+    })
+  })
+
+  it('resets form when discard button is clicked', () => {
+    const { changePasswordButton } = setupChangePasswordButton()
+
+    fireEvent.click(changePasswordButton)
+
+    const { passwordInput } = setupChangePasswordFields()
+    const discardButtonText = screen.getByText('common.cancel')
+
+    changeInputValue(passwordInput, 'oldPassword')
+
+    fireEvent.click(discardButtonText)
+    fireEvent.click(changePasswordButton)
+
+    const cleanedPasswordInput = screen.getByLabelText(
+      /editProfilePage.profile.passwordSecurityTab.currentPassword/i
+    )
+
+    expect(cleanedPasswordInput).toHaveValue('')
+  })
+
+  it('updates state when form fields are changed', () => {
+    const { changePasswordButton } = setupChangePasswordButton()
+
+    fireEvent.click(changePasswordButton)
+    const { currentPasswordInput } = setupChangePasswordFields()
+
+    changeInputValue(currentPasswordInput, 'oldPassword')
+
+    expect(currentPasswordInput).toHaveValue('oldPassword')
+  })
+})
